Fixes checkbox control when users uncheck the item
When user unchecks the checkbox the value isn't sent to server, making
it impossible to have all checkboxes unchecked.
This adds hidden field with value of nil **before** the checkbox. If
user selects the checkbox his choice will overwrite hidden field. If
user unchecks the checkbox his choice **will not be sent**, the hidden
field will be sent instead, simulating *unchecked* value.

See detailed explanation in Rails docs:
http://apidock.com/rails/ActionView/Helpers/FormHelper/check_box

This fixes bug that we have in community:
https://trello.com/c/i3vcHj6o/242-removing-badges-borks